+++ assets/javascripts/kitten/components/buttons/button.test.js
@@ -16,6 +16,7 @@ describe('<Button />', () => {
     expect(component).to.have.className('k-Button')
     expect(component).to.have.className('k-Button--hydrogen')
     expect(component).to.not.have.className('k-Button--helium')
+    expect(component).to.have.attr('disabled')
   })
 
   describe('name', () => {
@@ -45,5 +46,22 @@ describe('<Button />', () => {
     it('has an icon right', () => {
       expect(componentWithIconRight).to.have.className('k-Button--iconRight')
     })
+  })
+
+  describe('with default props', () => {
+    const defaultComponent = shallow(<Button />)
+
+    it('has a default name', () => {
+      expect(defaultComponent).attr('name', 'hydrogen')
+    })
+
+    it('has a default text', () => {
+      expect(defaultComponent).attr('text', 'Button')
+    })
+
+    it('has a default tag', () => {
+      expect(defaultComponent).attr('tag', 'button')
+    })
+
   })
 })